+++ db/migrate/20151207214812_add_position_to_admin_clinical_trials.rb
@@ -1,0 +1,5 @@
+class AddPositionToAdminClinicalTrials < ActiveRecord::Migration
+  def change
+    add_column :admin_clinical_trials, :position, :integer, null: false, default: 0
+  end
+end